+++ src/scripts/parser_peopleAtLocation.py
@@ -1,5 +1,5 @@
 import json
-import parsercache as cache
+import parser_cache as cache
 
 def printPeopleAtLocation(dictPerson, dictOrga, dictLocation):
     dictRes = {}
